pyproject.toml
[project]
name = "domaincheck"
version = "0.1.0"
description = "Domain lookup CLI tool with parallel execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.8.0",
    "python-whois>=0.9.6",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[project.scripts]
domaincheck = "main:main"
